+++ app/Http/Controllers/Backend/JptController.php
@@ -128,11 +128,15 @@ class JptController extends Controller
      * @param  \App\Models\Jpt  $jpt
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->authorize('view', Jpt::findOrFail($id));
-        Jpt::destroy($id);
-        return redirect()->route('jpt.index')->with('success', 'Data berhasil dihapus');
+        if ($request->delete == 'true') {
+            $this->authorize('view', Jpt::findOrFail($id));
+            Jpt::destroy($id);
+            return redirect()->route('jpt.index')->with('success', 'Data berhasil dihapus');
+        }
+        alert()->error('Gagal', 'Data gagal dihapus');
+        return redirect()->route('jpt.index');
     }
 
     public function createBongkar()
