+++ app/Http/Controllers/Backend/PelnasController.php
@@ -130,11 +130,15 @@ class PelnasController extends Controller
      * @param  \App\Models\Pelnas  $pelnas
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->authorize('view', Pelnas::findOrFail($id));
-        Pelnas::destroy($id);
-        return redirect()->route('pelnas.index')->with('success', 'Data berhasil dihapus');
+        if ($request->delete == 'true') {
+            $this->authorize('view', Pelnas::findOrFail($id));
+            Pelnas::destroy($id);
+            return redirect()->route('pelnas.index')->with('success', 'Data berhasil dihapus');
+        }
+        alert()->error('Gagal', 'Data gagal dihapus');
+        return redirect()->route('pelnas.index');
     }
 
     public function createBerangkat()
